// src/app/model/book-dto.ts
export class BookDto {
    authorKey: string[];
    authorName: string[];
    firstPublishYear: number;
    isbn: string[];
    key: string;
    // characters
    person: string[];
    personKey: string[];
    place: string[];
    placeKey: string[];
    ratingsAverage: number;
    ratingsCount: number;
    // tags kinda
    subject: string[];
    subjectKey: string[];
    title: string;
}